+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 2023-XX-XX 2.1.3
 * [MODDICONV-346](https://issues.folio.org/browse/MODDICONV-346) Fix unlinking for action-mapping associations
+* [MODDICONV-353](https://issues.folio.org/browse/MODDICONV-353) Add quickMARC - Default Create authority job profile
 
 ## 2023-10-12 2.1.0
 * [MODDICONV-345](https://issues.folio.org/browse/MODDICONV-345) Set wrapper ids at jsonb for migrated relations